+++ src/main/java/org/hanihome/hanihomebe/property/web/dto/request/create/PropertyCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package org.hanihome.hanihomebe.property.web.dto.request;
+package org.hanihome.hanihomebe.property.web.dto.request.create;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
@@ -6,7 +6,6 @@ import org.hanihome.hanihomebe.global.exception.CustomException;
 import org.hanihome.hanihomebe.global.response.domain.ServiceCode;
 import org.hanihome.hanihomebe.interest.region.Region;
 import org.hanihome.hanihomebe.property.domain.enums.GenderPreference;
-import org.hanihome.hanihomebe.property.domain.enums.ParkingOption;
 import org.hanihome.hanihomebe.property.domain.enums.PropertySuperType;
 import org.hanihome.hanihomebe.property.domain.vo.*;
 
@@ -45,8 +44,6 @@ public sealed interface PropertyCreateRequestDTO permits
 
 //    Set<LocalDateTime> availableFrom();
     MoveInInfo moveInInfo();
-
-    ParkingOption parkingOption();
 
     LocalDate meetingDateFrom();
     LocalDate meetingDateTo();
