+++ src/main/java/org/hanihome/hanihomebe/property/web/dto/request/patch/RentPropertyPatchRequestDTO.java
@@ -1,36 +1,39 @@
-package org.hanihome.hanihomebe.property.web.dto.request;
+package org.hanihome.hanihomebe.property.web.dto.request.patch;
 
 import jakarta.validation.constraints.AssertTrue;
 import lombok.Getter;
 import org.hanihome.hanihomebe.property.application.TimeSlotValidator;
-import org.hanihome.hanihomebe.property.domain.vo.ShareInternalDetails;
+import org.hanihome.hanihomebe.property.domain.vo.RentInternalDetails;
 import org.hanihome.hanihomebe.property.domain.command.PropertyPatchCommand;
-import org.hanihome.hanihomebe.property.domain.command.SharePropertyPatchCommand;
-import org.hanihome.hanihomebe.property.domain.enums.CapacityShare;
-import org.hanihome.hanihomebe.property.domain.enums.SharePropertySubType;
+import org.hanihome.hanihomebe.property.domain.command.RentPropertyPatchCommand;
+import org.hanihome.hanihomebe.property.domain.enums.CapacityRent;
+import org.hanihome.hanihomebe.property.domain.enums.RentPropertySubType;
 import org.hanihome.hanihomebe.property.domain.item.PropertyOptionItem;
 
 import java.util.List;
 
 @Getter
-public class SharePropertyPatchRequestDTO extends PropertyPatchRequestDTO {
-    private SharePropertySubType sharePropertySubType;   // 1. 매물 유형 (세컨드룸/마스터룸/거실쉐어)
-    private ShareInternalDetails internalDetails;
-    private CapacityShare capacityShare;                // 3. 수용 인원
+public class RentPropertyPatchRequestDTO extends PropertyPatchRequestDTO {
+    private RentPropertySubType rentPropertySubType;    // (RentProperty 고유) 매물 유형
+    private RentInternalDetails internalDetails;
+    private CapacityRent capacityRent;                  // (RentProperty 고유) 수용인원-렌트
 
     @AssertTrue(message = "timeSlot의 timeFrom, timeTo는 30분 단위여야 합니다.")
     private boolean isValidTimeSlot() {
-        // timeSlot is NULL
-        if (super.getMeetingDateFrom() == null && super.getMeetingDateTo() == null && super.getTimeSlots() == null) {
+        if (timeSlotIsNull()) {
             return true;    // patch라서 통과
         } else {
             return TimeSlotValidator.validateAllConditions(super.getTimeSlots());
         }
     }
 
+    private boolean timeSlotIsNull() {
+        return super.getMeetingDateFrom() == null && super.getMeetingDateTo() == null && super.getTimeSlots() == null;
+    }
+
     @Override
     public PropertyPatchCommand toCommand(List<PropertyOptionItem> propertyOptionItems) {
-        return SharePropertyPatchCommand.builder()
+        return RentPropertyPatchCommand.builder()
                 // 공통 필드
                 .genderPreference(super.getGenderPreference())
                 .lgbtAvailable(super.getLgbtAvailable())
@@ -40,19 +43,17 @@ public class SharePropertyPatchRequestDTO extends PropertyPatchRequestDTO {
                 .costDetails(super.getCostDetails())
                 .livingConditions(super.getLivingConditions())
                 .moveInInfo(super.getMoveInInfo())
-                .parkingOption(super.getParkingOption())
                 .meetingDateFrom(super.getMeetingDateFrom())
                 .meetingDateTo(super.getMeetingDateTo())
                 .timeSlots(super.getTimeSlots())
                 .viewingAlwaysAvailable(super.getViewingAlwaysAvailable())
                 .description(super.getDescription())
                 .displayStatus(super.getDisplayStatus())
-                .tradeStatus(super.getTradeStatus())
 
-                // ShareProperty 전용 필드
-                .sharePropertySubType(this.sharePropertySubType)
+                // RentProperty 전용 필드
+                .rentPropertySubType(this.rentPropertySubType)
                 .internalDetails(this.internalDetails)
-                .capacityShare(this.capacityShare)
+                .capacityRent(this.capacityRent)
                 .build();
     }
